Oswaal to NCERT
Step1: NCERT Social Science (four text books ) pdf is obtained.
Step2: Each of the pdf is converted into text and placed as four units (based on oswaal categorisation)
Step3: Topics associated with each unit is placed in the corresponding sub directory (unit).
Step 4: script to read and do semantic search and compare based on cosine similarity. - > prints the q, A, and top 5 matches.

Next: need to output the data as a common format.